requirements.txt
langchain==1.2.13
langchain-openai==1.1.11
langchain-community==0.4.1
langchain-ollama==1.0.1
langchain-chroma==1.1.0
chromadb==1.5.5
gradio==6.9.0
python-dotenv==1.1.0
pypdf==6.9.1
docx2txt==0.9
tiktoken==0.12.0
pydantic==2.12.5
ollama==0.6.1
